Outsources InOutSampler to its own file
