+++ src/axom/quest/CMakeLists.txt
@@ -142,7 +142,9 @@ if(AXOM_ENABLE_KLEE AND AXOM_ENABLE_SIDRE AND ((MFEM_FOUND AND AXOM_ENABLE_MFEM_
 endif()
 
 if(AXOM_ENABLE_KLEE AND AXOM_ENABLE_SIDRE AND (MFEM_FOUND AND AXOM_ENABLE_MFEM_SIDRE_DATACOLLECTION))
-    list(APPEND quest_headers SamplingShaper.hpp detail/shaping/shaping_helpers.hpp)
+    list(APPEND quest_headers SamplingShaper.hpp 
+                              detail/shaping/shaping_helpers.hpp
+                              detail/shaping/InOutSampler.hpp)
     list(APPEND quest_sources detail/shaping/shaping_helpers.cpp)
 endif()
